+++ R/model_diagnostics.R
@@ -14,6 +14,7 @@
 #' @references Explanatory Model Analysis. Explore, Explain, and Examine Predictive Models. \url{https://pbiecek.github.io/ema/}
 #' @export
 #' @examples
+#' library(DALEX)
 #' apartments_lm_model <- lm(m2.price ~ ., data = apartments)
 #' explainer_lm <- explain(apartments_lm_model,
 #'                          data = apartments,
@@ -45,9 +46,9 @@ model_diagnostics <-  function(explainer, variables = NULL, ...) {
   # if variables = NULL then all variables are added
   # otherwise only selected
   if (is.null(variables)) {
-    results <- explainer$data
+    results <- as.data.frame(explainer$data)
   } else {
-    results <- explainer$data[, intersect(variables, colnames(explainer$data)), drop = FALSE]
+    results <- as.data.frame(explainer$data[, intersect(variables, colnames(explainer$data)), drop = FALSE])
   }
   # is there target
   if (!is.null(explainer$y)) {
@@ -80,6 +81,7 @@ model_diagnostics <-  function(explainer, variables = NULL, ...) {
       results$residuals <- explainer$residuals[, 1] # this will work only for first column
     }
   }
+
   results$abs_residuals <- abs(results$residuals)
   results$label <- explainer$label
   results$ids <- seq_along(results$label)